Extract query builder into separate pkg
